refactor: . Move tests/*Renderer* to tests/Renderer/
