+++ tests/Renderer/TaskFieldRenderer.test.ts
@@ -2,8 +2,8 @@
  * @jest-environment jsdom
  */
 import moment from 'moment';
-import { TaskFieldHTMLData, TaskFieldRenderer } from '../src/Renderer/TaskFieldRenderer';
-import { TaskBuilder } from './TestingTools/TaskBuilder';
+import { TaskFieldHTMLData, TaskFieldRenderer } from '../../src/Renderer/TaskFieldRenderer';
+import { TaskBuilder } from '../TestingTools/TaskBuilder';
 
 window.moment = moment;
