+++ tests/Renderer/TaskLineRenderer.test.Visualise_HTML_Minimal_task_-_short_mode.approved.html
